% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPMM_get_project_info.R
\name{RPMM_get_project_info}
\alias{RPMM_get_project_info}
\title{get Project information of RIKEN PMM with Project-ID}
\usage{
RPMM_get_project_info(projectid = "RPMM0001")
}
\arguments{
\item{projectid}{RPMM Project ID (default: RPMM0001)}
}
\value{
A data.frame of Project info
}
\description{
Download project information from RIKEN PMM
(http://metabobank.riken.jp/) with SPARQL queries
}
\examples{
res <- RPMM_get_project_info(projectid = "RPMM0001") 
head(res)
}
\author{
Atsushi Fukushima
}
