% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPMM_get_projects_list.R
\name{RPMM_get_projects_list}
\alias{RPMM_get_projects_list}
\title{get Project list of RIKEN PMM with species info}
\usage{
RPMM_get_projects_list(sciname = "Arabidopsis thaliana", taxonid = 3702)
}
\arguments{
\item{sciname}{a scientific name of species (default: Arabidopsis thaliana)}

\item{taxonid}{NCBI taxonomy ID (default: 3702)}
}
\value{
A data.frame of Project lists
}
\description{
Download list of projects from RIKEN PMM 
(http://metabobank.riken.jp/) with SPARQL queries
}
\examples{
res <- RPMM_get_projects_list(taxonid = 3702) ## Arabidopsis
head(res)
}
\author{
Atsushi Fukushima
}
