% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPMM_get_rawdata_files.R
\name{RPMM_get_rawdata_files}
\alias{RPMM_get_rawdata_files}
\title{Getting all the raw data files in a Project from RPMM}
\usage{
RPMM_get_rawdata_files(project = "RPMM0001")
}
\arguments{
\item{project}{a Project identifier in RIKEN PMM (default: RPMM0001)}
}
\value{
A data.frame of raw data file lists
}
\description{
This function gets URIs of raw data files (e.g., cdf format) 
from RIKEN PMM 
(http://metabobank.riken.jp/) with SPARQL queries
}
\examples{
res <- RPMM_get_rawdata_files(project = 'RPMM0001')
head(res$url)
}
\author{
Atsushi Fukushima
}
