% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPMM_numOfprojects_family.R
\name{RPMM_numOfprojects_family}
\alias{RPMM_numOfprojects_family}
\title{Number Of Projects in RIKEN PMM with family info}
\usage{
RPMM_numOfprojects_family()
}
\value{
A data.frame of Project lists
}
\description{
This function counts projects from RIKEN PMM 
(http://metabobank.riken.jp/) for each family with SPARQL queries
}
\examples{
res <- RPMM_numOfprojects_family()
head(res)
}
\author{
Atsushi Fukushima
}
