% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPMM_numOfprojects_species.R
\name{RPMM_numOfprojects_species}
\alias{RPMM_numOfprojects_species}
\title{Number of Projects with species info}
\usage{
RPMM_numOfprojects_species()
}
\value{
A data.frame of Project lists
}
\description{
This function counts projects for each species from RIKEN PMM 
(http://metabobank.riken.jp/) with SPARQL queries
}
\examples{
res <- RPMM_numOfprojects_species()
head(res)
}
\author{
Atsu
}
